Add a new powerhint mode to boost GPU in shots on P22.
In P22 (Pro), there is no scaling_min_compute_freq (b/228921300) and under STREAMING_STANDARD setting GPU min frequency is 302000 which is smaller than scaling_min_compute_freq 471000 on P21 (Whi).

In addition, on Pro GPU we expect 25~30% slow down in the same frequency compared to Whi.

To boost the performance of GPU for shots on Pro, we add a new powerhint mode GCA_CAMERA_SHOT_GPU_EXTREME which raises the GPU frequency to the maximum.

Bug: 230289913.

Test: Pushed to local C10 (EVT1.0, go/ab/8471717) and test with cl/444178649 (trace: https://ui.perfetto.dev/#!/?s=23af98f7bc1729faefb9df2c6890566d48111bc957592a7c5ffbdad87c3d2f2d)
Change-Id: Ib77bddbdac32e689120c513f38029be9eab96449