Limit camera daemon cpuset to little cores when app launch
When exiting camera app, camera daemon is still busy in background,
which may affect the launch animation of launcher. So, limit the cpuset of camera daemon to little cores for 1 second when there is app launch hint to improve it, but cancel it right away when there is camera launch hint. Also bypass this behavior when video recording is active.

Bug: 240889279
Test: camera daemon cpuset changed as expected
Change-Id: Ieeafb0d96fbce1b7a97df0e554bd2223529bf1ed