Increase MIF frequency for 30fps (C10, P10), and 60fps (P10)
This CL increases MIF frequency for C10, P10 device in 30fps mode in order to
improve CPI, as well as P10 in 60fps mode (C10 for 60FPS was previously
addressed in ag/19263388). Although the MIF power consumption increases due to this
change, but the reduction in CPU cycles results in over-all SoC power saving.
This change also doesn't negatively impact framedrop rate.

Bug: 239240751
Test: ODPM and framedrop tests
Change-Id: I0454e25bd8687b6378016f14990935ff03e7a38b