Disable PMU at the start of camera launch
A recent change (ag/20028064) has caused regression in GCA launch
time due to activation of PMU at the camera launch time, and capping the
CPU frequency. This CL prevents PMU activation for as long as we are in
powerhints or power-boosts related to camera launch (LAUNCH, CAMERA_LAUNCH,
CAMERA_LAUNCH_EXTENDED).

Bug: 248718485
Test: Perfetto trace
Change-Id: I6d88e4ccaf6790578223202ee69c105360cfc559